Добавлены тесты всех методов контроллеров Faculty и Student

+++ src/main/java/bip/online/homework111352/conroller/FacultyController.java
@@ -4,6 +4,7 @@ import bip.online.homework111352.model.Faculty;
 import bip.online.homework111352.service.FacultyService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,10 +29,10 @@ public class FacultyController {
 
     @GetMapping("/search")
     public ResponseEntity<Collection<Faculty>> findFaculties(@RequestParam(required = false) String color, @RequestParam(required = false) String name) {
-        if (color != null && !color.isBlank()){
+        if (color != null && !color.isBlank()) {
             return ResponseEntity.ok(service.findByColor(color));
         }
-        if(name != null && !name.isBlank()) {
+        if (name != null && !name.isBlank()) {
             return ResponseEntity.ok(service.findByName(name));
         }
         return ResponseEntity.ok(Collections.emptyList());
@@ -55,7 +56,12 @@ public class FacultyController {
     )
     @GetMapping
     public ResponseEntity<Faculty> get(@RequestParam Long id) {
-        return ResponseEntity.ok(service.findById(id).orElseThrow());
+        if (service.findById(id).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            return ResponseEntity.ok(service.findById(id).get());
+        }
+
     }
 
     @Operation(
@@ -64,8 +70,12 @@ public class FacultyController {
     )
     @DeleteMapping
     public ResponseEntity delete(@RequestParam Long id) {
-        service.delete(id);
-        return ResponseEntity.ok().build();
+        if (service.findById(id).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            service.delete(id);
+            return ResponseEntity.ok().build();
+        }
     }
 
     @Operation(
